Offer request forms - refactor
